Use pipenv instead of pip + virtualenv
The old `requirements.txt` setup did not result in 100% deterministic
builds, as certain key dependencies were not pinned (and said project's
downstream dependencies, by design, are often not pinned).

This was okay for a long time, as there are few key dependencies, and
the software is designed to work with nearly any version of most
dependencies. Versions that are especially important (for example, the
version of Django) have long been pinned.

As the system grows more complex and more people rely on it, it's best
to start moving towards reproducible builds.